refactor: remove duplicated code in lambdas
