+++ modules/integrations/splunk_aws_billing/lambda/handler_per_service.py
@@ -2,118 +2,90 @@
 import calendar
 import io
 import json
-import os
-import re
 import time
 from decimal import ROUND_HALF_UP, Decimal
 from urllib.parse import unquote
 
 import boto3
+import common
 import pandas as pd
-import requests
-
-SPLUNK_HEC_URL = os.environ['SPLUNK_HEC_URL']
-SPLUNK_HEC_TOKEN = os.environ['SPLUNK_HEC_TOKEN']
-SPLUNK_INDEX = os.environ.get('SPLUNK_INDEX')
-SPLUNK_METRICS_TOKEN = os.environ['SPLUNK_METRICS_TOKEN']
-SPLUNK_METRICS_URL = os.environ['SPLUNK_METRICS_URL']
 
 s3 = boto3.client('s3')
 
 
-def send_to_splunk(event):
-    headers = {
-        'Authorization': f'Splunk {SPLUNK_HEC_TOKEN}',
-    }
-    try:
-        resp = requests.post(SPLUNK_HEC_URL, json=event,
-                             headers=headers, timeout=10)
-        print(
-            f'[Splunk Response] Status: {resp.status_code}, Body: {resp.text}')
-        resp.raise_for_status()
-    except requests.HTTPError as e:
-        print(f'[HTTPError] {e.response.status_code} - {e.response.text}')
-    except requests.RequestException as e:
-        print(f'[RequestException] Error sending to Splunk: {e}')
+def create_event_body(row, fields):
+    usage_date = str(row['usage_date'])
+    event_time = calendar.timegm(pd.to_datetime(usage_date).timetuple())
 
-
-def send_metric_to_o11y(metric_name, value, timestamp, dimensions):
-    headers = {
-        'X-SF-TOKEN': SPLUNK_METRICS_TOKEN,
-        'Content-Type': 'application/json'
-    }
-    payload = {
-        'gauge': [{
-            'metric': metric_name,
-            'value': value,
-            'timestamp': timestamp * 1000,
-            'dimensions': dimensions
-        }]
-    }
-    try:
-        resp = requests.post(SPLUNK_METRICS_URL,
-                             headers=headers, json=payload, timeout=5)
-        print(
-            f'[O11y] Metric {metric_name} sent: {resp.status_code} {resp.text}')
-        resp.raise_for_status()
-    except requests.RequestException as e:
-        print(f'[O11y ERROR] Failed to send metric {metric_name}: {e}')
-
-
-def extract_arn_parts(arn):
-    match = re.search(
-        r'arn:aws:resource-groups:(?P<aws_region>[\w-]+):(?P<account_id>\d+):group/(?P<forgecicd_tenant>[^-]+)-(?P<forgecicd_region_alias>[^-]+)-(?P<forgecicd_vpc_alias>[^/]+)/',
-        arn
-    )
-    if match:
-        return match.groupdict()
     return {
-        'aws_region': 'unknown',
-        'account_id': 'unknown',
-        'forgecicd_tenant': 'unknown',
-        'forgecicd_region_alias': 'unknown',
-        'forgecicd_vpc_alias': 'unknown'
+        'source': 'aws-cur-per-service',
+        'sourcetype': 'forgecicd:aws:billing:cur',
+        'index': common.SPLUNK_INDEX,
+        'event': {
+            'service': row['line_item_product_code'],
+            'aws_application': row['user_aws_application'],
+            'cost_usd': float(Decimal(row['line_item_unblended_cost']).quantize(Decimal('0.00001'), rounding=ROUND_HALF_UP)),
+            'net_cost_usd': float(Decimal(row['line_item_net_unblended_cost']).quantize(Decimal('0.00001'), rounding=ROUND_HALF_UP)),
+            'usage_date': usage_date,
+            'event_time': event_time,
+            **fields
+        }
     }
 
 
-def preprocess_df(df):
-    print(f'[INFO] Raw DataFrame shape: {df.shape}')
-    df['line_item_usage_start_date'] = pd.to_datetime(
-        df['line_item_usage_start_date'])
-    df['ingest_time'] = pd.Timestamp.now(tz='UTC')
+def send_batches(batch, metrics_batch):
+    if batch:
+        common.send_to_splunk_batch(batch)
+    if metrics_batch:
+        common.send_metric_to_o11y_batch(metrics_batch)
 
-    def parse_tags(val):
-        if isinstance(val, list):
-            try:
-                return dict(val)
-            except (TypeError, ValueError):
-                return {}
-        elif isinstance(val, dict):
-            return val
-        elif isinstance(val, str):
-            try:
-                return json.loads(val)
-            except json.JSONDecodeError:
-                return {}
-        return {}
 
-    df['resource_tags'] = df['resource_tags'].apply(parse_tags)
-    df['user_aws_application'] = df['resource_tags'].apply(
-        lambda tags: tags.get('user_aws_application', 'unknown'))
-    df = df[df['user_aws_application'] != 'unknown']
-    df['usage_date'] = df['line_item_usage_start_date'].dt.date
+def process_grouped_rows(grouped, now_ts):
+    batch = []
+    current_size = 0
+    metrics_batch = []
 
-    key_cols = [
-        'line_item_usage_start_date',
-        'line_item_product_code',
-        'user_aws_application',
-        'identity_line_item_id'
-    ]
-    df = df.sort_values('ingest_time', ascending=False)
-    df = df.drop_duplicates(subset=key_cols, keep='first')
+    for _, row in grouped.iterrows():
+        fields = common.extract_arn_parts(row['user_aws_application'])
+        event_body = create_event_body(row, fields)
+        line = json.dumps(event_body)
+        line_size = len(line.encode())
 
-    print(f'[INFO] Preprocessed DataFrame shape: {df.shape}')
-    return df
+        # Flush logs batch if limits reached
+        if len(batch) >= common.MAX_BATCH_COUNT or current_size + line_size > common.MAX_BATCH_SIZE_BYTES:
+            common.send_to_splunk_batch(batch)
+            batch = []
+            current_size = 0
+
+        batch.append(line)
+        current_size += line_size
+
+        # Prepare metrics
+        dimensions = {
+            'usage_date': str(row['usage_date']),
+            'service': row['line_item_product_code'],
+            'aws_application': row['user_aws_application'],
+            **fields
+        }
+        metrics_batch.append({
+            'metric': 'forge.per_service.cost_usd',
+            'value': event_body['event']['cost_usd'],
+            'timestamp': now_ts,
+            'dimensions': dimensions
+        })
+        metrics_batch.append({
+            'metric': 'forge.per_service.net_cost_usd',
+            'value': event_body['event']['net_cost_usd'],
+            'timestamp': now_ts,
+            'dimensions': dimensions
+        })
+
+        # Flush metrics batch if big enough
+        if len(metrics_batch) >= common.METRICS_BATCH_SIZE:
+            common.send_metric_to_o11y_batch(metrics_batch)
+            metrics_batch = []
+
+    send_batches(batch, metrics_batch)
 
 
 def lambda_handler(event, context):
@@ -122,12 +94,12 @@ def lambda_handler(event, context):
     for record in event['Records']:
         bucket = record['s3']['bucket']['name']
         key = unquote(record['s3']['object']['key'])
-
         print(f'[INFO] Processing file from bucket: {bucket}, key: {key}')
+
         obj = s3.get_object(Bucket=bucket, Key=key)
         df = pd.read_parquet(io.BytesIO(obj['Body'].read()))
 
-        df = preprocess_df(df)
+        df = common.preprocess_df(df)
 
         grouped = df.groupby(
             ['usage_date', 'line_item_product_code', 'user_aws_application'],
@@ -137,47 +109,11 @@ def lambda_handler(event, context):
             'line_item_net_unblended_cost': 'sum'
         })
 
-        now_ts = int(time.time())
-
         print(f'[INFO] Grouped {len(grouped)} records for Splunk')
 
-        for _, row in grouped.iterrows():
-            fields = extract_arn_parts(row['user_aws_application'])
-            usage_date = str(row['usage_date'])
-            event_time = calendar.timegm(
-                pd.to_datetime(usage_date).timetuple())
+        now_ts = int(time.time())
 
-            event_data = {
-                'source': 'aws-cur-per-service',
-                'sourcetype': 'forgecicd:aws:billing:cur',
-                'index': SPLUNK_INDEX,
-                'event': {
-                    'service': row['line_item_product_code'],
-                    'aws_application': row['user_aws_application'],
-                    'cost_usd': float(Decimal(row['line_item_unblended_cost']).quantize(Decimal('0.00001'), rounding=ROUND_HALF_UP)),
-                    'net_cost_usd': float(Decimal(row['line_item_net_unblended_cost']).quantize(Decimal('0.00001'), rounding=ROUND_HALF_UP)),
-                    'usage_date': usage_date,
-                    'event_time': event_time,
-                    **fields
-                }
-            }
-
-            print(f'[DEBUG] Sending to HEC: {json.dumps(event_data)}')
-            send_to_splunk(event_data)
-
-            dimensions = {
-                'usage_date': usage_date,
-                'service': row['line_item_product_code'],
-                'aws_application': row['user_aws_application'],
-                **fields
-            }
-
-            print(f'[DEBUG] Sending to O11y: {json.dumps(dimensions)}')
-
-            send_metric_to_o11y(
-                'forge.per_service.cost_usd', event_data['event']['cost_usd'], now_ts, dimensions)
-            send_metric_to_o11y(
-                'forge.per_service.net_cost_usd', event_data['event']['net_cost_usd'], now_ts, dimensions)
+        process_grouped_rows(grouped, now_ts)
 
     print('[INFO] Lambda execution finished.')
     return {'statusCode': 200}
